Split action creators into individual files
